+++ clients/pkg/promtail/config/config.go
@@ -7,9 +7,6 @@ import (
 	"github.com/go-kit/log"
 	"github.com/go-kit/log/level"
 	dskit_flagext "github.com/grafana/dskit/flagext"
-
-	"github.com/grafana/loki/pkg/tracing"
-
 	"gopkg.in/yaml.v2"
 
 	"github.com/grafana/loki/clients/pkg/promtail/client"
@@ -18,7 +15,8 @@ import (
 	"github.com/grafana/loki/clients/pkg/promtail/scrapeconfig"
 	"github.com/grafana/loki/clients/pkg/promtail/server"
 	"github.com/grafana/loki/clients/pkg/promtail/targets/file"
-
+	"github.com/grafana/loki/clients/pkg/promtail/wal"
+	"github.com/grafana/loki/pkg/tracing"
 	"github.com/grafana/loki/pkg/util/flagext"
 )
 
@@ -39,6 +37,7 @@ type Config struct {
 	LimitsConfig    limit.Config          `yaml:"limits_config,omitempty"`
 	Options         Options               `yaml:"options,omitempty"`
 	Tracing         tracing.Config        `yaml:"tracing"`
+	WAL             wal.Config            `yaml:"wal"`
 }
 
 // RegisterFlags with prefix registers flags where every name is prefixed by