+++ pkg/usagestats/reporter.go
@@ -221,9 +221,6 @@ func (rep *Reporter) readSeedFile(ctx context.Context) (*ClusterSeed, error) {
 	if err != nil {
 		return nil, err
 	}
-	if err != nil {
-		return nil, err
-	}
 	defer func() {
 		if err := reader.Close(); err != nil {
 			level.Error(rep.logger).Log("msg", "failed to close reader", "err", err)